05_Output/Projects/@Active/Kyoritsu-Seiyaku-DX/poc-handwriting-extraction/requirements.txt
# PDF処理
pdf2image>=1.16.3
PyMuPDF>=1.23.0

# 画像処理
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# OCR (スタブ用、後でCloud Vision等に差し替え)
# google-cloud-vision>=3.5.0
# google-cloud-aiplatform>=1.38.0

# CLI
click>=8.1.0

# 型ヒント
typing-extensions>=4.8.0
